I forgot to commit the relevant changes on the script, so I'll use this commit to explain the whole program, while I copy this in the README file.
- First, I created the function that prompts the user
to enter a choice (rock paper or scissors). It didn't
matter if the user input in lower case or capital letters,
I added the toLowerCase() function to turn everything the same
case to compare later.

- Then, created the function that creates the computer's
choice, by selecting a random number, and dividing the
possible choices in 3, so if Math.random returns a
value from 0 to 1 (1 / 3 = 0,333 aprox.). So every
value that's returned, with if conditionals, whether it
is under 0.33, 0.66 or 0.99, the function gives the
computer a random choice.

- After both choices are set, then the game starts.
The only thing that the game function does is to
compare the values of each choice, and, depending
on which choice is the winner one, it will add a point
to the winner.

- Then, the last function is created to repeat the game
5 times with a while loop, so that the game does not stop
until 5 rounds are played. One the loop finishes, the
function checks which of the two scores is higher and
chooses a winner. If there's a tie, the function
returns it.